Add files via upload
Word Guessing Games as the name implies is a game where the user guesses a random word. The user has only one chance to guess the right word. In a situation where the user is unable to guess correctly the word is given to the user and the user has to re-run the game and a new word is given.  The first and the last letter of the word as a hint. The user is also given the number of words in the letter.