Add files via upload
This program is a simple calculator that performs the 4 basic mathematical operations which are :
Addition, Subtraction, Multiplication and Division. Each operation has a number so in order for the user to perform a caluculation , the user has to input the number of the desired mathematical operation and then input the numbers to be calculatedand separate them by leaving a single space in-between.